First version of "fragment block" to unify register and stack models
A fragment block is a memory-like block of fragments of values. It provides
for reading and writing values. There are also lower-level functions for
reading and writing lists of "bytes" (memvals).

The blocks are conceptually infinite and indexed by "offsets" expressed in
words (32-bit blocks). Accesses touch a 32- or 64-bit quantity. This limits
the number of cases of overlapping accesses to take into account.

Using quantities also maps cleanly to encoding values using Many32/Many64
chunks. This way, values are encoded using the projection/injection
functions that can produce Fragments or Undefs, but never Byte memvals. This
avoids some kinds of type punning that can be modeled by CompCert for
general memory accesses (encoding a value as Bytes, then decoding a
subsequence of those Bytes to a different value).

Fragment blocks are meant to be the single unified basis for modeling the
register file at Mach and Asm levels, as well as stack frames. In
particular, the byte copy functions provide for spilling 64-bit registers
without having to know if the fragments stored within encode a single 64-bit
value or two 32-bit values.