Preserve storage class for functions declared within a block
This is a follow-up to the introduction of the Storage_auto storage class
in commit 760e422.

For functions declared within a block, we used to set their storage class
to "extern":

{                          {
  int f(void);       --->     extern int f(void);
}                          }

This helped enter_or_refine_ident understand that those declarations
have linkage.

Now that we have Storage_auto, this commit teaches enter_or_refine_ident
that local declarations have no linkage if auto or static, and have
linkage if extern or default.

Then, there is no need to change storage class to extern for
locally-declared functions.

In turn, this improves the "extern-after-definition" warning recently
introduced, avoiding a bizarre warning in the following case:

int foo(void){return 0;}
int main(void){
    int foo(void);
    return foo();
}