Parameterize elab_expr by the full elaboration context
Before, we would pass just the `ctx_vararg` component of the context
to `elab_expr` as a Boolean argument.

For future extensions, we will need to pass more of the context to
`elab_expr`, so why not pass the whole context?

This is what this commit does.  The `stmt_context` type is renamed
`elab_context` and its definition is moved earlier.  The `ctx`
argument is passed as is from `elab_stmt` to `elab_expr`.