Preserve attribute(("aligned")) in the AST, don't map it to _Alignas
We used to recognize attribute(("aligned"(N))) and map it to _Alignas(N)
during elaboration.

However, we want to restrict the places where _Alignas can occur, as
standardized in ISO C11, while leaving more freedom for the placement
of the "aligned" attribute.

As a first step in this direction, this commit keeps the "aligned"
attribute unchanged in the AST, and distinct from _Alignas attributes.
Both attributes are honored when it comes to determining the actual
alignment of a type.