Improve reporting of configuration errors
The full "usage" text is so long that the actual error message scrolls off
the screen.

With this commit, instead, a short (3-line) error message is printed,
with a suggestion to do `./configure -help`.

The whole "usage" text is printed by `./configure -help`.

Also: better error message for unknown options (arguments starting with `-`).