Attach _Alignas to names and refactor _Alignas checks (#133)
* Refactor common code of alignas.

Instead of working on attributes the function now works directly
on the type since the check always performed an extraction of
attributes from a type. 
Bug 23393

* Attach _Alignas to the name.

Bug 23393

* Attach "aligned" attributes to names

So that __attribute((aligned(N))) remains consistent with _Alignas(N).
gcc and clang apply "aligned" attributes to names, with a special case
for typedefs:
    typedef __attribute((aligned(16))) int int_al_16;
    int_al_16 * p;
    __attribute((aligned(16))) int * q;

For gcc, p is naturally-aligned pointer to 16-aligned int and
q is 16-aligned pointer to naturally-aligned int.

For CompCert with this commit, both p and q are 16-aligned pointers
to naturally-aligned int.

* Resurrect the alignment test involving typedef

The test was removed because it involved an _Alignas in a typedef,
which is no longer supported.  However the same effect can be achieved
with an "aligned" attribute, which is still supported in typedef.