x86: wrong modeling of ZF flag for FP comparisons
As written in the comment, ZF should be set if the two floats are
equal or unordered.  The "or unordered" case was missing in the
original modeling of FP comparisons.

- Set ZF flag correctly in the Asm.compare_floats and Asm.compare_floats32 functions.
- Update the proofs in Asmgenproof1 accordingly.

No change required to the code generated for FP comparisons: this code
already anticipated the "or unordered" case.

Problem reported by Alix Trieu.