<stddef.h>: define NULL with type void *
ISO C2011 7.19 para 3 says "NULL, which expands to an implementation-defined null pointer constant"

ISO C2011 6.3.2.3 para 3 says "An integer constant expression with the value 0, or such an expression cast to type void *, is called a null pointer constant."

So, it seems NULL can be defined either as "0" or as "(void *) 0".  However, the two definitions are not equivalent in the context of a call to a variadic or unprototyped function: passing 0 when the function expects a pointer value can go wrong if sizeof(int) != sizeof(void *).

This commit changes the definition of NULL to the safer one:

This is what GCC and Clang do in C mode; they use "#define NULL 0" for C++ only.

Fixes issue #265