Distinguish object-related and name-related attributes
This is a second step towards mimicking GCC/Clang's handling of attributes.

This commit introduces a distinction between
- Object-related attributes, such as "section", which apply to the object (function, variable) being defined;
- Name-related attributes, such as "aligned", which apply to the name (object, struct/union member, struct/union/enum tag) being defined.

In particular, "aligned" is now attached to "struct" and "union" definitions, while it used to be "floated up" before.

The C11 _Alignas modifier is treated like an object-related attribute, so that
```
       struct s { ... };
       _Alignas(64) struct s x;
```
correctly associates the alignment with "x" and not with "struct s", where it would be ignored because it was not part of the original definition of s.