Reject object-related and struct-related attributes on typedefs
This commit adds a check to reject type definitions such as
```
typedef __attribute((section "foo")) int fooint;
```
GCC and Clang also reject this as an error.

Without the check, the behavior is somewhat surprising:
```
fooint x;    // placed in section "foo"
fooint * x;  // placed in default section, attribute "foo" is ignored
```

Note that the following must be accepted:
```
typedef struct { ... } __attribute((packed)) t;
```
The "packed" attribute is correctly attached to the struct type and should not be checked.  This is achieved by using `attribute_of_type_no_expand` to get the attributes of the typedef-ed type, excluding the attributes carried by a struct/union or another typedef.