Reworked range entries.
The fist changes changes the offset for range entries to used labels
instead of integer constants, leaving the computation to the
assembler.

The second part of the change the address changes the way ranges
entries of scopes are printed. They need to be relative to the start
address of the code in the section they are included.

Bug 26234