Asmgenproof0: add predicate exec_straight0
This is a variant of exec_straight where it is allowed to take zero steps.
In other words, exec_straight0 is the "star" relation, while exec_straight
is the "plus" relation.
In the end we need "plus" relations in simulation diagrams, to show
the absence of stuttering.  But the "star" relation exec_straight0 is
useful to reason about code fragments that are always preceded or
followed by at least one instruction.