Change the expected types for arguments to __builtin_annot, and extended asm
Currently, the arguments to __builtin_annot, __builtin_ais_annot,
__builtin_debug, and extended asm statements are treated like
arguments to an unprototyped or vararg function call.  In particular,
arguments of type "float" are converted to "double", generating useless
code.

To avoid this extra, useless conversion, this commit changes the types
expected for the arguments to these built-ins and to extended asm
statements.  Now they are the types of the arguments themselves, after
performing the usual unary conversions (e.g. char -> int), but without
the problematic float -> double conversion.  This ensures that no code
is generated to change the representation of the arguments.