Added new diagnostic for non-linear conditionals
The new diagnostics is triggered if a conditional is used that
may not be transformed into linear code by the later by the
if conversion.

The new diagnostic is emitted if a conditional may contain an
unsafe expression or is contained within another conditional,
logical and or logical or expression. An expression is unsafe if
it contains a call, changes memory or if its evaluation leads to
undefined behavior, for example division and modulo.

Also fixes a small typo in a comment in Cutil.