Coq 8.10 compatibility: (temporarily) silence new warning
The "undeclared-scope" warning fires when we use a "notation" scope
before having declared it.  This is a good thing, except that
the "Declare Scope" vernacular that declares a scope was introduced
in Coq 8.10 and is not available in earlier versions.

Hence there is no way to avoid triggering the warning yet remain
compatible with pre-8.10 Coq versions.

This commit silences the warning.  It will have to revisited when
Coq 8.10 is the oldest version of Coq we support in CompCert.