clightgen -dclight: print function parameters correctly
The Clight output of clightgen is Clight version 2, after SimplLocals
conversion, where function parameters are temporary variables, not
variables.

This commit makes sure the function parameters are printed as
temporary variables and not as variables.  In passing, it
generalizes the Clight pretty-printer so that it can print
both Clight version 1 and Clight version 2.

Closes: #314