-dclight output: use nicer names for temporary variables
The temporary variables introduced by SimplLocals reuse the same
integer identifiers as the local variables they come from.  This commit
ensures that these variables are printed as "$var", where "var"
is the original variable name, instead of "$NNN" as before.

The "$NNN" form is retained for temporary variables that do not
correspond to a source-level local variable, such as the temporary
variables introduced by SimplExpr.

This commit should make no difference for "ccomp -dclight", because
the Clight that is printed is the Clight version 1 produced by
SimplExpr, where every temporary is fresh and does not correspond
to a source-level local variable.

This commit does change the output of "clightgen -dclight", because
the Clight that is printed is the Clight version 2 produced by
SimplLocals.  The printed Clight is much more legible thanks to
the more meaningful temporary variable names.