Fix for AArch64 alignment problem (#206)
In addressing modes for load and store instructions, the offset must be a multiple of the memory size being accessed.  When accessing global variables, this may not be the case if the alignment of the variable is less than its size.  Errors occur at link time.

This PR extends the check for a representable offset for the addressing of global
variables to also check whether the variable is correctly aligned.  Only if both conditions are
met can we generate the short sequence Padrp / ADadr.  Otherwise we go through the generic
loadsymbol sequence.