Revised menhirLib autoconfiguration
Since Menhir version 20200123, we need to link with menhirLib.cmxa
instead of menhirLib.cmx.

This commit chooses automatically the file to link with:
menhirLib.cmxa if it exists in the menhirLib installation directory,
menhirLib.cmx otherwise.

To reliably find the installation directory, configure was changed
to record the menhirLib directory in Makefile.config, variable MENHIR_DIR,
instead of a pre-cooked command-line option MENHIR_INCLUDES.

Makefile.extr was adapted accordingly.

Fixes: #329
Closes: #330