Define IRC.class_of_type for types Tany32, Tany64
Until now the types Tany32 and Tany64 were not used prior to register
allocation, so IRC.class_of_type did not need to be defined for
those types.

However, there are possible uses of stack slots of type Tany32 and Tany64
to specify calling conventions.  For this purpose, we need to define
class_of_type for Tany32 and Tany64.  We follow the informal convention
that Tany32 goes in integer registers and Tany64 goes into integer
registers if 64-bit wide, or FP registers otherwise.