Eliminate known builtins whose result is ignored
A typical example is `(void) __builtin_sel(a, b, c)`.

It is safe to generate zero code for these uses of builtins
because builtins whose semantics are known to the compiler
are pure.  Other builtins with side effects (e.g. `__builtin_trap`)
are not known and will remain in the compiled code.

It is useful to generate zero code for these uses of builtins
because some of them (e.g. `__builtin_sel`) must be transformed
into proper CminorSel expressions during instruction selection.
Otherwise, they propagate all the way to ExpandAsm, causing
a "not implemented" error there.