Fix regression on PowerPC / Diab
On PowerPC/Diab, common declarations must not be used for small data sections.

Add a `~common` option to `PrintAsmaux.variable_section` to control
the use of common declarations.  The default is whatever is specified
on the command line using the `-fcommon` and `-fno-common` options.

Use `~common:false` for `Section_small_data` on PowerPC / Diab.

Note that on PowerPC/Linux, GCC uses common declarations for uninitialized
variables in small data section, so we keep doing this in CompCert as well.