Selectively turn some Coq 8.14 warnings off
Warning "deprecated hint rewrite without locality" cannot be addressed:
the suggested fix (qualify with Global or Local or [#export])
is not supported by Coq 8.11 to 8.13 !

Warning "deprecated instance without locality" is turned off
for generated file cparser/Parser.v only.

Menhir needs to be changed so as to generate the `Global` modifier
that would silence this warning.