Resurrect a warning for bit fields of enum types
Earlier CompCert versions would warn if a bit field of width N and
type enum E was too small for the values of the enumeration: whether
the field is interpreted as a N-bit signed integer or a N-bit unsigned
integer, some values of the enumeration are not representable.

This warning was performed in the Bitfields emulation pass, which went
away during the reimplementation of bit fields within the verified
part of CompCert.

In this commit, we resurrect the warning and perform it during the Elab pass.

In passing, some of the code that elaborates bit fields was moved to a
separate function "check_bitfield".