Add explicit tail calls to Clight and to C#minor
By popular demand.

The tail call optimization pass performed later in CompCert will not
turn a call into a tail call if the function has stack-allocated data,
because this might change the semantics.

Having an explicit Stailcall form in Clight makes it possible to force
some calls in tail position to be tail calls, deallocating. deallocate
stack-allocated data before the call.  This is the same reason why we
have an Stailcall form in Cminor already.