Enforce evaluation order in IRC.add_interf and IRC.add_pref
As written previously, OCaml can evaluate `nodeOfVar g v1` and
`nodeOfVar g v2` in any order.  Consequently, `v1` and `v2` can be
entered in the `varTable` hash table in different order.  This can
result in different (but equally valid) register allocations.

It's better to enforce one evaluation order for the sake of
reproducible compilations when the OCaml version changes.