Introduce float_conversion_default_nan parameter for float-float conversions
For RISC-V we need to return the canonical NaN value if the argument of
a float32->float64 or float64->float32 conversion is any NaN.
This is in line with 11.3 from the RISC-V manual, the description of
the conversion operations as well as what the spike ISA simulator and
qemu do.

Other platforms convert the NaN payload (by truncation or expansion)
in float32->float64 and float64->float32 conversions.

Fixes: #428