Support mergeable sections for fixed-size literals
On platforms that support them (ELF, macOS), use mergeable sections
(like `.rodata.cst8`) for 4-, 8- and 16-byte wide literals.

Works only if the LITERAL section is the default one.  If the user
provided their own LITERAL section, all literals are put in it
regardless of their sizes.

Support for mergeable string sections is introduced in this commit too
but needs further changes in C2C.ml .