Remove unused functions.
The `reset_constants` was only used for ARM and there it was superfluous
since the jumptables, which it additionally resets in contrast to
`reset_literals`, are always empty since jumptables are handled directly
in during printing of the instruction.

The `exists_constants` function is now also unused since we test directly
for the different sizes in `emit_constants`.