Finer treatment of temporaries destroyed by built-in functions (part 1)
`destroyed_by_builtin` now returns two lists of registers:
- The first list is the registers destroyed "before" the builtin,
  i.e. before reading the arguments of the builtin.  Any register in
  this first list cannot be allocated to any of the parameters.
- Second list is the temporaries destroyed "during" the builtin, i.e.
  after reading the arguments, but before writing the results.
  The registers in this second list can be allocated to any of the parameters.

This enables some simplifications in the expansion of built-in functions
in $ARCH/Asmexpand.ml.  In particular, it is no longer necessary to
choose a temporary (among a list of temporaries destroyed "during")
that is not used by any of the parameters: it suffices to declare one
temporary as destroyed "before".

This commit updates the semantics of LTL, Linear and Mach to reflect
the effect of the "destroyed before" temporaries.

The register allocator and its validator are updated to take
"destroyed before" temporaries into account.

Proofs of semantic preservation were updated accordingly.

Only x86 is handled at this point.  Other target architectures remain to be
updated.